move packages by team into own class
